consolidate selector tests
