+++ test/data/mock-state.json
@@ -109,7 +109,7 @@
           "chainId": "4",
           "isEns": false,
           "memo": "",
-          "name": ""
+          "name": "Address Book Account 1"
         }
       }
     },
